feat: allow query flag with data exploration insight flag (#14845)
