+++ cypress/productAnalytics/index.ts
@@ -83,8 +83,14 @@ export const insight = {
     newInsight: (insightType: string = 'TRENDS'): void => {
         const networkInterceptAlias = interceptInsightLoad(insightType)
 
-        cy.get('[data-attr=menu-item-insight]').click() // Open the new insight menu in the sidebar
-        cy.get(`[data-attr="sidebar-new-insights-overlay"][data-attr-insight-type="${insightType}"]`).click()
+        if (insightType === 'JSON') {
+            cy.get('[data-attr=menu-item-insight]').click() // Open the new insight menu in the sidebar
+            cy.get('[data-attr="sidebar-new-insights-overlay"][data-attr-insight-type="TRENDS"]').click()
+            insight.clickTab('JSON')
+        } else {
+            cy.get('[data-attr=menu-item-insight]').click() // Open the new insight menu in the sidebar
+            cy.get(`[data-attr="sidebar-new-insights-overlay"][data-attr-insight-type="${insightType}"]`).click()
+        }
 
         cy.wait(`@${networkInterceptAlias}`)
     },
