+++ frontend/src/scenes/insights/InsightNav/insightNavLogic.tsx
@@ -1,4 +1,4 @@
-import { actions, connect, kea, key, listeners, path, props, selectors } from 'kea'
+import { actions, connect, kea, key, listeners, path, props, reducers, selectors } from 'kea'
 import { InsightLogicProps, InsightType } from '~/types'
 
 import type { insightNavLogicType } from './insightNavLogicType'
@@ -10,7 +10,7 @@ import { insightDataLogic, queryFromKind } from 'scenes/insights/insightDataLogi
 import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
 import { FEATURE_FLAGS } from 'lib/constants'
 import { insightMap } from '~/queries/nodes/InsightQuery/utils/queryNodeToFilter'
-import { isDataTableNode, isHogQLQuery, isInsightVizNode } from '~/queries/utils'
+import { containsHogQLQuery, isInsightVizNode } from '~/queries/utils'
 import { examples, TotalEventsTable } from '~/queries/examples'
 import { LemonTag } from 'lib/lemon-ui/LemonTag/LemonTag'
 
@@ -38,32 +38,43 @@ export const insightNavLogic = kea<insightNavLogicType>([
     actions({
         setActiveView: (view: InsightType) => ({ view }),
     }),
+    reducers({
+        userSelectedView: {
+            setActiveView: (_, { view }) => view,
+        },
+    }),
     selectors({
         isUsingDataExploration: [
             (s) => [s.featureFlags],
             (featureFlags) => !!featureFlags[FEATURE_FLAGS.DATA_EXPLORATION_INSIGHTS],
         ],
         allowQueryTab: [
-            (s) => [s.featureFlags, s.isUsingDataExploration],
-            (featureFlags, isUsingDataExploration) =>
-                isUsingDataExploration && !!featureFlags[FEATURE_FLAGS.DATA_EXPLORATION_QUERY_TAB],
+            (s) => [s.featureFlags],
+            (featureFlags) => !!featureFlags[FEATURE_FLAGS.DATA_EXPLORATION_QUERY_TAB],
         ],
         activeView: [
-            (s) => [s.filters, s.query, s.isUsingDataExploration],
-            (filters, query, isUsingDataExploration) => {
-                if (isUsingDataExploration) {
-                    if (isInsightVizNode(query)) {
-                        return insightMap[query.source.kind] || InsightType.TRENDS
-                    } else if (!!query) {
-                        if (isHogQLQuery(query) || (isDataTableNode(query) && isHogQLQuery(query.source))) {
+            (s) => [s.filters, s.query, s.userSelectedView],
+            (filters, query, userSelectedView) => {
+                // if userSelectedView is null then we must be loading an insight
+                // and, we can prefer a present query over a present filter
+                // otherwise we can have both a filter and a query and without userSelectedView we don't know which to use
+                // so, if there is a user selected view, we use that
+                // this gets much simpler once everything is using queries
+
+                if (userSelectedView === null) {
+                    if (!!query) {
+                        if (containsHogQLQuery(query)) {
                             return InsightType.SQL
+                        } else if (isInsightVizNode(query)) {
+                            return insightMap[query.source.kind] || InsightType.TRENDS
+                        } else {
+                            return InsightType.JSON
                         }
-                        return InsightType.JSON
                     } else {
-                        return InsightType.TRENDS
+                        return filters.insight || InsightType.TRENDS
                     }
                 } else {
-                    return filters.insight || InsightType.TRENDS
+                    return userSelectedView
                 }
             },
         ],
@@ -137,32 +148,40 @@ export const insightNavLogic = kea<insightNavLogicType>([
     }),
     listeners(({ values, actions }) => ({
         setActiveView: ({ view }) => {
-            if (values.isUsingDataExploration) {
-                if (view === InsightType.TRENDS) {
-                    actions.setQuery(queryFromKind(NodeKind.TrendsQuery))
-                } else if (view === InsightType.FUNNELS) {
-                    actions.setQuery(queryFromKind(NodeKind.FunnelsQuery))
-                } else if (view === InsightType.RETENTION) {
-                    actions.setQuery(queryFromKind(NodeKind.RetentionQuery))
-                } else if (view === InsightType.PATHS) {
-                    actions.setQuery(queryFromKind(NodeKind.PathsQuery))
-                } else if (view === InsightType.STICKINESS) {
-                    actions.setQuery(queryFromKind(NodeKind.StickinessQuery))
-                } else if (view === InsightType.LIFECYCLE) {
-                    actions.setQuery(queryFromKind(NodeKind.LifecycleQuery))
-                } else if (view === InsightType.JSON) {
+            if ([InsightType.SQL, InsightType.JSON].includes(view as InsightType)) {
+                // if the selected view is SQL or JSON then we must have the "allow queries" flag on,
+                // so no need to check it
+                if (view === InsightType.JSON) {
                     actions.setQuery(TotalEventsTable)
                 } else if (view === InsightType.SQL) {
                     actions.setQuery(examples.HogQLTable)
                 }
             } else {
-                actions.setFilters(
-                    cleanFilters(
-                        // double-check that the view is valid
-                        { ...values.filters, insight: view === InsightType.JSON ? InsightType.TRENDS : view },
-                        values.filters
+                if (values.isUsingDataExploration) {
+                    if (view === InsightType.TRENDS) {
+                        actions.setQuery(queryFromKind(NodeKind.TrendsQuery))
+                    } else if (view === InsightType.FUNNELS) {
+                        actions.setQuery(queryFromKind(NodeKind.FunnelsQuery))
+                    } else if (view === InsightType.RETENTION) {
+                        actions.setQuery(queryFromKind(NodeKind.RetentionQuery))
+                    } else if (view === InsightType.PATHS) {
+                        actions.setQuery(queryFromKind(NodeKind.PathsQuery))
+                    } else if (view === InsightType.STICKINESS) {
+                        actions.setQuery(queryFromKind(NodeKind.StickinessQuery))
+                    } else if (view === InsightType.LIFECYCLE) {
+                        actions.setQuery(queryFromKind(NodeKind.LifecycleQuery))
+                    }
+                } else {
+                    actions.setFilters(
+                        cleanFilters(
+                            // double-check that the view is valid
+                            { ...values.filters, insight: view || InsightType.TRENDS },
+                            values.filters
+                        ),
+                        undefined,
+                        true
                     )
-                )
+                }
             }
         },
     })),
