+++ frontend/src/scenes/insights/insightDataLogic.ts
@@ -8,7 +8,6 @@ import { insightLogic } from './insightLogic'
 import { queryNodeToFilter } from '~/queries/nodes/InsightQuery/utils/queryNodeToFilter'
 import { filtersToQueryNode } from '~/queries/nodes/InsightQuery/utils/filtersToQueryNode'
 import { isInsightVizNode } from '~/queries/utils'
-import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
 import { cleanFilters } from './utils/cleanFilters'
 import { nodeKindToDefaultQuery } from '~/queries/nodes/InsightQuery/defaults'
 import { dataNodeLogic } from '~/queries/nodes/DataNode/dataNodeLogic'
@@ -44,9 +43,7 @@ export const insightDataLogic = kea<insightDataLogicType>([
     connect((props: InsightLogicProps) => ({
         values: [
             insightLogic,
-            ['insight', 'isUsingDataExploration'],
-            featureFlagLogic,
-            ['featureFlags'],
+            ['insight', 'isUsingDataExploration', 'isUsingDashboardQueries'],
             // TODO: need to pass empty query here, as otherwise dataNodeLogic will throw
             dataNodeLogic({ key: insightVizDataNodeKey(props), query: {} as DataNode }),
             ['dataLoading as insightDataLoading', 'responseErrorObject as insightDataError'],
@@ -137,26 +134,28 @@ export const insightDataLogic = kea<insightDataLogicType>([
             }
         },
         saveInsight: ({ redirectToViewMode }) => {
-            if (!values.query) {
-                return
-            }
-
             let filters = values.insight.filters
-            if (isInsightVizNode(values.query)) {
+            if (values.isUsingDataExploration && isInsightVizNode(values.query)) {
                 const querySource = values.query.source
                 filters = queryNodeToFilter(querySource)
-            } else if (values.isQueryBasedInsight) {
+            } else if (values.isUsingDashboardQueries && values.isQueryBasedInsight) {
                 filters = {}
+            }
+
+            let query = undefined
+            if (values.isUsingDashboardQueries && values.isQueryBasedInsight) {
+                query = values.query
             }
 
             actions.setInsight(
                 {
                     ...values.insight,
                     filters: filters,
-                    ...(values.isQueryBasedInsight ? { query: values.query } : {}),
+                    query: query ?? undefined,
                 },
                 { overrideFilter: true, fromPersistentApi: false }
             )
+
             actions.insightLogicSaveInsight(redirectToViewMode)
         },
     })),
