+++ frontend/src/scenes/saved-insights/newInsightsMenu.tsx
@@ -18,12 +18,13 @@ const insightTypeURL: Record<InsightType, string> = {
 }
 
 function insightTypesForMenu(isUsingDataExplorationQueryTab: boolean): [string, InsightTypeMetadata][] {
-    let menuEntries = Object.entries(INSIGHT_TYPES_METADATA)
+    // never show JSON InsightType in the menu
+    let menuEntries = Object.entries(INSIGHT_TYPES_METADATA).filter(([insightType]) => insightType !== InsightType.JSON)
+
     if (!isUsingDataExplorationQueryTab) {
-        menuEntries = menuEntries.filter(
-            ([insightType]) => insightType !== InsightType.JSON && insightType !== InsightType.SQL
-        )
+        menuEntries = menuEntries.filter(([insightType]) => insightType !== InsightType.SQL)
     }
+
     return menuEntries
 }
